+++ RSNNS/src/kr_amap.cpp
@@ -627,7 +627,7 @@ int   SnnsCLib::kram_getClassNo (void)
     if ((i > ArtMap_NoOfRecUnits_b) && ((*topo_ptr)->act < 0.9)) {
        return (-1);
     } else {
-       return (topo_ptr - ArtMap_map_layer + 1);
+       return (int) (topo_ptr - ArtMap_map_layer + 1);
     } /*if*/
 
 } /* kram_getClassNo () */
